move channels to new file channels.ts to modularize backend code
